Create a base op
